Wording/typo fixes
I've gone through up to Types and fixed a bunch of awkward phrasings, typos, misplaced punctuation marks, etc. Probably made it a bit more formal than the original text though so please let me know if you want me to de-formalize them.

Also unrelated, but I want to know what are your thoughts about italicizing words kept in english, as that (I believe) is the EYD specification for it? Will be a lot of work though, but been there done during skripsi days so I'm happy to work on that.